:recycle: api: move ComplementaryCertificationScoringWithoutComplementaryReferential to its domain
